src/requirements.txt
certifi==2019.11.28
chardet==3.0.4
hupper==1.9.1
idna==2.8
PasteDeploy==2.0.1
plaster==1.0
plaster-pastedeploy==0.7
pyramid==1.10.4
python-decouple==3.3
requests==2.22.0
translationstring==1.3
urllib3==1.25.7
venusian==3.0.0
waitress==1.4.2
WebOb==1.8.5
zope.deprecation==4.4.0
zope.interface==4.7.1
